change unsigned to uint8_t for SCSI structure

+++ Common/DtaStructures.h
@@ -415,56 +415,56 @@ public:
         OPCODE  = 0x12,
     };
     uint8_t         m_Opcode;                               //  0
-    unsigned        m_EVPD                  : 1;            //  1
-    unsigned        m_Reserved_1            : 7;
+    uint8_t        m_EVPD                  : 1;            //  1
+    uint8_t        m_Reserved_1            : 7;
     uint8_t         m_PageCode;                             //  2
     uint16_t        m_AllocationLength;                     //  3
     uint8_t         m_Control;                              //  5
 } ;                                  //  6
-/*
+
 ////////////////////////////////////////////////////////////////////////////////
 class CScsiCmdInquiry_StandardData
 ////////////////////////////////////////////////////////////////////////////////
 {
 public:
-    unsigned        m_PeripheralDeviceType      : 5;        //  0
-    unsigned        m_PeripheralQualifier       : 3;
-    unsigned        m_Reserved_1                : 6;        //  1
-    unsigned        m_LUCong                    : 1;
-    unsigned        m_RMB                       : 1;
+    uint8_t        m_PeripheralDeviceType      : 5;        //  0
+    uint8_t        m_PeripheralQualifier       : 3;
+    uint8_t        m_Reserved_1                : 6;        //  1
+    uint8_t        m_LUCong                    : 1;
+    uint8_t        m_RMB                       : 1;
     uint8_t         m_Version;                              //  2
-    unsigned        m_ResponseDataFormat        : 4;        //  3
-    unsigned        m_HiSup                     : 1;
-    unsigned        m_NormACA                   : 1;
-    unsigned        m_Reserved_2                : 1;
-    unsigned        m_Reserved_3                : 1;
+    uint8_t        m_ResponseDataFormat        : 4;        //  3
+    uint8_t        m_HiSup                     : 1;
+    uint8_t        m_NormACA                   : 1;
+    uint8_t        m_Reserved_2                : 1;
+    uint8_t        m_Reserved_3                : 1;
     uint8_t         m_AdditionalLength;                     //  4
-    unsigned        m_Protect                   : 1;        //  5
-    unsigned        m_Reserved_4                : 2;
-    unsigned        m_3PC                       : 1;
-    unsigned        m_TPGS                      : 2;
-    unsigned        m_ACC                       : 1;
-    unsigned        m_SCCS                      : 1;
-    unsigned        m_ADDR16                    : 1;        //  6
-    unsigned        m_Reserved_5                : 2;
-    unsigned        m_Obsolete_1                : 1;
-    unsigned        m_MultiP                    : 1;
-    unsigned        m_VS1                       : 1;
-    unsigned        m_EncServ                   : 1;
-    unsigned        m_Obsolete_2                : 1;
-    unsigned        m_VS2                       : 1;        //  7
-    unsigned        m_CmdQue                    : 1;
-    unsigned        m_Reserved_6                : 1;
-    unsigned        m_Obsolete_3                : 1;
-    unsigned        m_Sync                      : 1;
-    unsigned        m_WBus16                    : 1;
-    unsigned        m_Reserved_7                : 1;
-    unsigned        m_Obsolete_4                : 1;
+    uint8_t        m_Protect                   : 1;        //  5
+    uint8_t        m_Reserved_4                : 2;
+    uint8_t        m_3PC                       : 1;
+    uint8_t        m_TPGS                      : 2;
+    uint8_t        m_ACC                       : 1;
+    uint8_t        m_SCCS                      : 1;
+    uint8_t        m_ADDR16                    : 1;        //  6
+    uint8_t        m_Reserved_5                : 2;
+    uint8_t        m_Obsolete_1                : 1;
+    uint8_t        m_MultiP                    : 1;
+    uint8_t        m_VS1                       : 1;
+    uint8_t        m_EncServ                   : 1;
+    uint8_t        m_Obsolete_2                : 1;
+    uint8_t        m_VS2                       : 1;        //  7
+    uint8_t        m_CmdQue                    : 1;
+    uint8_t        m_Reserved_6                : 1;
+    uint8_t        m_Obsolete_3                : 1;
+    uint8_t        m_Sync                      : 1;
+    uint8_t        m_WBus16                    : 1;
+    uint8_t        m_Reserved_7                : 1;
+    uint8_t        m_Obsolete_4                : 1;
     uint8_t         m_T10VendorId[8];                       //  8
     uint8_t         m_ProductId[16];                        // 16
     uint8_t         m_ProductRevisionLevel[4];              // 32
 };                                  // 36
-*/
+
 
 ////////////////////////////////////////////////////////////////////////////////
 class CScsiCmdSecurityProtocolIn
@@ -478,8 +478,8 @@ public:
     uint8_t         m_Opcode;                   //  0
     uint8_t         m_SecurityProtocol;         //  1
     uint16_t        m_SecurityProtocolSpecific; //  2
-    unsigned        m_Reserved_1    : 7;        //  4
-    unsigned        m_INC_512       : 1;
+    uint8_t        m_Reserved_1    : 7;        //  4
+    uint8_t        m_INC_512       : 1;
     uint8_t         m_Reserved_2;               //  5
     uint32_t        m_AllocationLength;         //  6
     uint8_t         m_Reserved_3[1];            // 10
@@ -499,8 +499,8 @@ public:
     uint8_t         m_Opcode;                   //  0
     uint8_t         m_SecurityProtocol;         //  1
     uint16_t        m_SecurityProtocolSpecific; //  2
-    unsigned        m_Reserved_1    : 7;        //  4
-    unsigned        m_INC_512       : 1;
+    uint8_t        m_Reserved_1    : 7;        //  4
+    uint8_t        m_INC_512       : 1;
     uint8_t         m_Reserved_2;               //  5
     uint32_t        m_TransferLength;           //  6
     uint8_t         m_Reserved_3[1];            // 10
